Isolated Berkely code from the rest, renamed simulator.py to adapter.py.
