+++ pacman/behaviors.py
