+++ pacman/berkeley/README.md
@@ -1,0 +1,4 @@
+Berkeley Code
+=============
+
+This code is from [Project 5: Classification](http://ai.berkeley.edu/classification.html) v1.001. "Unused" files (in this project) were removed and [pacman.py](pacman.py) was changed only to **not** keep track of which game states have been explored .
