+++ pacman/berkeley/game.py
@@ -4,7 +4,7 @@
 # educational purposes provided that (1) you do not distribute or publish
 # solutions, (2) you retain this notice, and (3) you provide clear
 # attribution to UC Berkeley, including a link to http://ai.berkeley.edu.
-#
+# 
 # Attribution Information: The Pacman AI projects were developed at UC Berkeley.
 # The core projects and autograders were primarily created by John DeNero
 # (denero@cs.berkeley.edu) and Dan Klein (klein@cs.berkeley.edu).
@@ -29,7 +29,7 @@ import sys
 # Parts worth reading #
 #######################
 
-class Agent(object):
+class Agent:
     """
     An agent must define a getAction method, but may also define the
     following methods which will be called if they exist:
