+++ pacman/berkeley/util.py
