+++ pacman/layouts/classic1Ghost.lay
