+++ pacman/layouts/classic2Ghosts.lay
