+++ pacman/layouts/classic3Ghosts.lay
