+++ pacman/layouts/classic4Ghosts.lay
