+++ pacman/layouts/medium1Ghosts.lay
