+++ pacman/layouts/medium2Ghosts.lay
