+++ pacman/learn.py
