+++ pacman/learning.py
