+++ pacman/test
@@ -3,33 +3,32 @@
 # http://stackoverflow.com/questions/2694588/fork-two-processes-and-kill-the-second-when-the-first-is-done
 
 # Basic testing for all experiments (may take a while....)
-ADAPTER="simulator.py"
+ADAPTER="adapter.py"
 CONTROLLER="controller.py"
 AGENT_OPTIONS="--ghost-agent ai --pacman-agent ai"
 ITERATION_OPTIONS="--learn-num 2 --test-num 1"
-POLICY_OPTIONS="--policy-file $POLICY_FILE"
 
 OPTIONS="$AGENT_OPTIONS $ITERATION_OPTIONS"
 
 for EXP in "1" "2" "3" "4" "5" "6"
 do
-	echo -e "\n\nExperiment $EXP"
+	echo -e "\n\nExperiment $EXP\n"
 	python2 $CONTROLLER & pid_controller=$!
 	python2 $ADAPTER $OPTIONS -e $EXP & pid_adapter=$!
 	wait $pid_adapter
 	kill $pid_controller
 
-	echo -e "\n\nExperiment $EXP with policy --"
-	$POLICY_FILE="experiment_$EXP.pol"
+	echo -e "\n\nExperiment $EXP with new policies\n"
+	POLICY_FILE="experiment_$EXP.pol"
 	rm -f $POLICY_FILE
 	python2 $CONTROLLER & pid_controller=$!
-	python2 $ADAPTER $OPTIONS -e $EXP $POLICY_OPTIONS & pid_adapter=$!
+	python2 $ADAPTER $OPTIONS -e $EXP --policy-file $POLICY_FILE & pid_adapter=$!
 	wait $pid_adapter
 	kill $pid_controller
 
-	echo -e "\n\nExperiment $EXP test with existing Policy file"
+	echo -e "\n\nExperiment $EXP test with existing policy file\n"
 	python2 $CONTROLLER & pid_controller=$!
-	python2 $ADAPTER $OPTIONS -e $EXP $POLICY_OPTIONS & pid_adapter=$!
+	python2 $ADAPTER $OPTIONS -e $EXP --policy-file $POLICY_FILE & pid_adapter=$!
 	wait $pid_adapter
 	kill $pid_controller
 done
