+++ pacman/test.py
