+++ pacman/test_learn.py
