--- simulator/pacman_learn.py
@@ -1,7 +1,0 @@
-import learn
-
-class PacmanProblemAdapter(learn.ProblemAdapter):
-    pass
-
-class PacmanAgent(learn.Agent):
-    pass
